% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleEnrichment}
\alias{exampleEnrichment}
\title{Example Enrichment Results}
\format{
A data frame with 3 columns, source, pathway, and enrichment_score.
\describe{
  \item{source}{The source/group used when applying GSEA.}
  \item{pathway}{The pathway name when applying GSEA.}
  \item{enrichment_score}{Enrichment scores form GSEA results.}
}
}
\source{
A subset of public data.
}
\usage{
exampleEnrichment
}
\description{
A subset of the colorectal data where gene-set enrichment was done both for differentially expressed genes ("Expression") and differentially spliced genes ("Splicing").
}
\keyword{datasets}
