% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleGeneSets}
\alias{exampleGeneSets}
\title{Example Gene Sets}
\format{
A named list of gene sets, where the names represent the gene set
  names and each list element contains the corresponding genes for that
  gene set.
}
\source{
A subset of public data.
}
\usage{
exampleGeneSets
}
\description{
A named list of gene sets, where the names represent the gene set names and
each list element contains the corresponding genes for that gene set.
}
\keyword{datasets}
