% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleGsList}
\alias{exampleGsList}
\title{Example Gene Sets}
\format{
A gsList object with 204 gene sets from 2 source(s) clustered into 4 labeled clusters.
}
\usage{
exampleGsList
}
\description{
A gsList object with 204 gene sets from 2 source(s) clustered into 4 labeled clusters.
}
\keyword{datasets}
