% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractResolution}
\alias{extractResolution}
\title{Extract default resolution}
\usage{
extractResolution(geneSetsList)
}
\arguments{
\item{geneSetsList}{'gsList' object that has been clustered with the 'clusterGeneSets()' function.}
}
\value{
default resolution
}
\description{
Extract default resolution
}
\examples{
gsListObject <- updateResolution(exampleGsList, 1.1)
}
