% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsList.R
\docType{package}
\name{geneSetSimplifyR-package}
\alias{geneSetSimplifyR-package}
\title{geneSetSimplifyR: Simplify Gene-Set Enrichment Analysis}
\description{
geneSetSimplifyR seeks to simplify gene-set enrichment analysis (GSEA) via clustering of gene sets. Given a list of gene sets of interest the package clusters these using LSI and SNN and subsequently the clusters are labelled using TF-IDF to find the best suitable two-word label for each cluster. The package contains several plotting functions so the user can quickly visualize the clustering output and draw biological meaningful insights.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chihs-dtu/geneSetSimplifyR}
  \item Report bugs at \url{https://github.com/chihs-dtu/geneSetSimplifyR/issues}
}

}
\author{
\strong{Maintainer}: Kristoffer Vitting-Seerup \email{krivi@dtu.dk} (\href{https://orcid.org/0000-0002-6450-0608}{ORCID})

Authors:
\itemize{
  \item Ella Hedeboe \email{ellahedeboe@hotmail.com} (\href{https://orcid.org/0000-0002-3978-4661}{ORCID})
  \item Chiao-Yu Hsieh \email{chihs@dtu.dk} (\href{https://orcid.org/0009-0007-0463-5977}{ORCID})
}

}
\keyword{internal}
