% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeList.R
\name{initializeList}
\alias{initializeList}
\title{Initialize gene sets list}
\usage{
initializeList(geneSetsList, geneSetsDF = NULL, verbose = TRUE)
}
\arguments{
\item{geneSetsList}{A list containing the gene sets to be analyzed.
The name of each element in the list should be a gene set and each element should contain
the genes belonging to that gene set.}

\item{geneSetsDF}{Optional data.frame with gene sets and enrichment scores that can be usefull
for downstream visualization (both for effect size and direction and if multiple conditions are analyzed).
Should have a column named 'pathway' containing the gene sets and a column
named 'enrichment_score' or 'NES' with the enrichment of each gene set.
Optinally have a column named 'source' indicating the source of each gene set. This allows for
analysis of multi-condtional data throughput the entire workflow and is also supported by many
plotting functions. If a gene-set was found in multiple conditions there should be multiple rows
for the same 'pathway' indicating its respective source and corresponding enrichment_score.
See the vignette for examples.}

\item{verbose}{a logic indicating whether to print progress statements. Defaults to TRUE.}
}
\value{
'gsList' object with a gene sets dataframe and enrichment scores for each gene set.
}
\description{
Initialize a 'gsList' object from a dataframe with gene sets and enrichment scores and a gene sets database.
}
\examples{
gsListObject <- initializeList(geneSetsList = exampleGeneSets, geneSetsDF = exampleEnrichment)
}
