% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kvsElbow.R
\name{kvsElbow}
\alias{kvsElbow}
\title{Determine the elbow cutoff to be used to run UMAP}
\usage{
kvsElbow(variance, zScoreCutoff = 2.5)
}
\arguments{
\item{variance}{Stored in a Seurat object as 'seuratObj@reductions$lsi@stde'}

\item{zScoreCutoff}{Integer indicating the cut off value for the Z-score}
}
\value{
Elbow cutoff
}
\description{
Determine the elbow cutoff to be used to run UMAP
}
\examples{
elbowRes <- kvsElbow(variance = seuratObj@reductions$lsi@stdev)
}
