% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterLabels.R
\name{plotClusterLabels}
\alias{plotClusterLabels}
\title{Plot label information for a particular cluster}
\usage{
plotClusterLabels(
  geneSetsList,
  clusterNumber,
  resolution = NULL,
  nLabels = 5,
  maxChar = 50
)
}
\arguments{
\item{geneSetsList}{'gsList' object that has been clustered with the 'clusterGeneSets()' function.}

\item{clusterNumber}{Which cluster to plot. Here the cluster number is used}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{nLabels}{The number of labels to plot. Default to 5}

\item{maxChar}{The max number of characters of labels shown in plot. Longer labels will be truncated using "...". Defaults to 50}
}
\value{
a ggplot2 object.
}
\description{
Plot label information for a particular cluster
}
\examples{
gsListObject <- plotClusterLabels(geneSetsList = exampleGsList, clusterNumber = 0)
}
