% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustree.R
\name{plotClustree}
\alias{plotClustree}
\title{Plot cluster tree}
\usage{
plotClustree(
  geneSetsList,
  node_alpha = 1,
  node_text_size = 2.5,
  node_size_range = c(4, 20),
  edge_width = 0.7,
  node_text_angle = 30,
  resolutionsToPlot = NULL
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{node_alpha}{Numeric value indicating the alpha of nodes}

\item{node_size_range}{Numeric vector with two values indicating the
minimum and maximum point size of nodes, respectively}

\item{edge_width}{Numeric value indicating the width of plotted edges}

\item{node_text_angle}{Numeric value indicating the rotation of node labels}

\item{resolutionsToPlot}{Resolutions to include in plot. Use Null to plot all. Default is NULL}

\item{node_text_sixe}{Numeric value indicating node text size}
}
\value{
A ggplot2 object
}
\description{
Graphs a clustering tree plot showing t
}
\examples{
plotClustertree(geneSetsList)
}
