% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendogram.R
\name{plotDendogram}
\alias{plotDendogram}
\title{Plot dendogram of cluster averages}
\usage{
plotDendogram(geneSetsList, resolution = NULL, removeClusterId = FALSE)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}
}
\value{
A list of ggplot objects.
}
\description{
A plot showing the hierarchical clustering of the cluster averages
(dendogram). A tool made for showing how clusters (not individual gene-sets) relate to each other.
}
\examples{
plotProportions(geneSetsList)
}
