% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot enrichment as heatmap}
\usage{
plotHeatmap(
  geneSetsList,
  resolution = NULL,
  removeClusterId = FALSE,
  highligthClusterNo = NULL,
  transpose = FALSE,
  cutree_rows = NA,
  cutree_cols = NA,
  drawPlot = TRUE,
  ...
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}

\item{highligthClusterNo}{A vector with the cluster numbers of the cluster (at the resolution) to highligth in the umap}

\item{transpose}{A logic indicating whether to transpose the heatmap (swich rows and cols)}

\item{cutree_rows}{pheatmap control paramter. Number of clusters the rows are divided into}

\item{cutree_cols}{pheatmap control paramter. Number of clusters the columns are divided into}
}
\value{
A list of ggplot objects.
}
\description{
Creates a heatmap of showing mean enrichment each gene-set cluster in each sources (usefull for datasets with many data sources).
}
\examples{
plotEnrichment(geneSetsList)
}
