% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Scatter plot of UMAP dimensionality reduction}
\usage{
plotUMAP(
  geneSetsList,
  resolution = NULL,
  removeClusterId = FALSE,
  interactive = FALSE,
  highlightClusterNo = NULL,
  pointSize = 2,
  pointAlpha = 1/3,
  labelSize = 3,
  labelAlpha = 1,
  highlightPointAlpha = 2/3,
  highlightPointSize = 4,
  highlightLabelAlpha = 1,
  highlightLabelSize = 5
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}

\item{interactive}{Logical, indicating whether to plot an interactive UMAP plot. Default: FALSE.
If TRUE, an interactive plot will be plotted using the 'HoverLocator()' from the Seurat package.}

\item{highlightClusterNo}{A vector with the cluster numbers of the cluster (at the resolution) to highlight in the UMAP}

\item{pointSize}{size of points. Default is 2}

\item{pointAlpha}{transparency of points. Default is 1/3.}

\item{labelSize}{size of label Default is 2}

\item{labelAlpha}{transparency of label Default is 1.}

\item{highlightPointAlpha}{The transparency of the highlighted clusters. Set different from "alpha" to use transparency for highlighting clusters. Default is 2/3}

\item{highlightPointSize}{The size of the highlighted clusters. Set different from "size" to use size for highlighting clusters. Default is 2}

\item{highlightLabelAlpha}{The transparency of the highlighted clusters. Set different from "alpha" to use transparency for highlighting clusters. Default is 2/3}

\item{highlightLabelSize}{The size of the highlighted clusters. Set different from "size" to use size for highlighting clusters. Default is 2}
}
\value{
A list of ggplot objects.
}
\description{
Graphs the output of the UMAP dimensionality reduction as a scatterplot
using the 'DimPlot()' function from the Seurat package.
Gene sets are colored and labelled by their cluster.
When there is only one source in the data (default),
each point will be shaped by the direction of the enrichment score of the gene set.
When there are two sources in the data,
each point will be shaped by the source indicating whether the gene set was found
significant in either or both sources.
}
\examples{
plotUMAP(exampleGsList)
}
