% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Violin plot of clustering}
\usage{
plotViolin(
  geneSetsList,
  resolution = NULL,
  removeClusterId = FALSE,
  highlightClusterNo = NULL
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}

\item{highlightClusterNo}{A vector with the cluster numbers of the cluster (at the resolution) to highlight in the UMAP}
}
\value{
A list of ggplot objects.
}
\description{
Creates a violin plot of the enrichment scores for each cluster.
}
\examples{
plotViolin(geneSetsList = exampleGsList, highlightClusterNo = c(0,1), removeClusterId = FALSE)
}
