% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopWords}
\alias{stopWords}
\title{Stop Words}
\format{
A list of character strings.
}
\usage{
stopWords
}
\description{
A list of character strings used for annotating the cluster.
}
\keyword{datasets}
