% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateClusterLabel.R
\name{updateClusterLabels}
\alias{updateClusterLabels}
\title{Plot label information for a particular cluster}
\usage{
updateClusterLabels(
  geneSetsList,
  resolution = NULL,
  clusterNumber,
  newClusterName
)
}
\arguments{
\item{geneSetsList}{'gsList' object that has been clustered with the 'clusterGeneSets()' function.}

\item{resolution}{Float. The clustering resolution to use as the default.
The labels of this resolution will be set as the 'active.idents' in the Seurat object containing the clustering,
and thus this is the resolution used for the different plotting functions.
The 'resolution' must have been used as a resolution in the 'clusterGeneSets()' function.}

\item{clusterNumber}{Which cluster to plot. Here the cluster number is used}

\item{newClusterName}{The new name of the cluster}
}
\value{
'gsList' object with a single cluster label updated. The original names are also stored.
}
\description{
Plot label information for a particular cluster
}
\examples{
exampleGsList <- updateClusterLabels(geneSetsList = exampleGsList, 
clusterNumber = 0, newClusterName = 'cancer response')
}
