% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateResolution}
\alias{updateResolution}
\title{Update default resolution of gsList}
\usage{
updateResolution(geneSetsList, resolution)
}
\arguments{
\item{geneSetsList}{'gsList' object that has been clustered with the 'clusterGeneSets()' function.}

\item{resolution}{Float. The clustering resolution to use as the new default.
The new 'resolution' must have been used as a resolution in the 'clusterGeneSets()' function.}
}
\value{
'gsList' object updated with new default resolution
}
\description{
Update default resolution of gsList
}
\examples{
gsListObject <- updateResolution(exampleGsList, 1.1)
}
