% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGeneSets.R
\name{clusterGeneSets}
\alias{clusterGeneSets}
\title{Cluster gene sets}
\usage{
clusterGeneSets(
  geneSetsList,
  resolution = c(0.1, 0.3, 0.6, 1.1, 1.6, 2.1, 3.1, 4.1, 5.1),
  algorithm = "louvain",
  verbose = TRUE
)
}
\arguments{
\item{geneSetsList}{'gsList' object created with the 'initializeList()' function.}

\item{resolution}{Vector with the desired clustering resolutions to use. Higher number usually results in more clusters.}

\item{algorithm}{Algorithm for clustering. Options: 'louvain', 'leiden'}

\item{verbose}{a logic indicating whether to print progress statements. Defaults to TRUE.}
}
\value{
'gsList' object with clustering stored. The best resolution (highest
 silhouette score) is annoated and used throughout the rest of the workflow (unless
 specified in subsequent functions).
}
\description{
Cluster gene sets
}
\examples{
# Create gsListObject
geneSetsList <- initializeList(geneSetsList = exampleGeneSets, geneSetsDF = exampleEnrichment)
gsListObject <- clusterGeneSets(geneSetsList = geneSetsList)
}
