% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichments.R
\name{plotEnrichments}
\alias{plotEnrichments}
\title{Plot enrichment as lollipop plot}
\usage{
plotEnrichments(
  geneSetsList,
  resolution = NULL,
  removeClusterId = FALSE,
  highligthClusterNo = NULL
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}

\item{highligthClusterNo}{A vector with the cluster numbers of the cluster (at the resolution) to highligth in the umap}
}
\value{
A list of ggplot objects.
}
\description{
Creates a lollipop plot of showing mean enrichment and size of each gene-set cluster.
}
\examples{
plotEnrichments(exampleGsList)
}
