% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{plotProportions}
\alias{plotProportions}
\title{Plot proportions of gene sets in each cluster for each source}
\usage{
plotProportions(
  geneSetsList,
  resolution = NULL,
  removeClusterId = FALSE,
  highligthClusterNo = NULL
)
}
\arguments{
\item{geneSetsList}{'gsList' object with clusters and labels.}

\item{resolution}{The cluster resolution to plot. Default to NULL which means the optimal clustering found by clusterGeneSets().}

\item{removeClusterId}{Logical, if TRUE the cluster number will be removed from each cluster label. Default: FALSE.}

\item{highligthClusterNo}{A vector with the cluster numbers of the cluster (at the resolution) to highligth in the umap}
}
\value{
A list of ggplot objects.
}
\description{
Graphs a bar plot showing the proportion of gene sets within
each cluster that comes from each of the two sources in the analyzed data.
Gene-sets found in both sources contribute to the fraction of both sources.
}
\examples{
data("exampleGsList")
plotProportions(exampleGsList)
}
